requirements.txt
accelerate==0.34.2
av==13.0.0
certifi==2024.8.30
charset-normalizer==3.3.2
decord==0.6.0
einops==0.8.0
filelock==3.16.1
fsspec==2024.9.0
huggingface-hub==0.25.1
idna==3.10
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
packaging==24.1
pandas==2.2.3
pillow==10.4.0
psutil==6.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
qwen-vl-utils==0.0.8
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
sentencepiece==0.2.0
six==1.16.0
sympy==1.12
timm==1.0.9
tokenizers==0.20.0
torch==2.3.0+cu118
torchaudio==2.3.0+cu118
torchvision==0.18.0+cu118
tqdm==4.66.5
transformers==4.45.1
triton==2.3.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
